pyproject.toml
[project]
name = "understanding-langchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "chromadb>=0.5.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "numpy>=1.24.0"
]
